feat: Added pso mm moderation channel

+++ bot/src/events/guildMemberRemove.ts
@@ -5,7 +5,15 @@ import { teamService } from '../services/teamService';
 export default {
 	name: 'guildMemberRemove',
 	async execute(guildMember: GuildMember) {
+		const team = await teamService.findTeamByGuildId(guildMember.guild.id)
+		if (!team) {
+			return
+		}
 		await teamService.removeCaptain(guildMember.guild.id, guildMember.id)
 		await teamService.removePlayer(guildMember.guild.id, guildMember.id)
+
+		if (team.verified) {
+			await teamService.notifyNoLongerVerified(guildMember.client, team, `${guildMember.user} left the discord server.`)
+		}
 	}
 } as IEventHandler
